requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
aiosqlite>=0.20.0
pydantic>=2.9.2
pillow>=10.4.0
natsort>=8.4.0
sse_starlette
pydantic-multipart